+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-
 ### Release v1.0.4
 
 ### Feature
@@ -10,6 +9,7 @@
 - Remove "To" on contract creation on Eth Transfers on Address Transfers and Transactions Table [#903](https://github.com/EthVM/EthVM/pull/903)
 - Contract creation [#912](https://github.com/EthVM/EthVM/pull/912)
 - Enforce branch names [#911](https://github.com/EthVM/EthVM/pull/911)
+- Create workflow for master and staging [#916](https://github.com/EthVM/EthVM/pull/916)
 
 ===================================================================================